Implement all the changes required by the previous code reviewer
