+++ src/components/Categories.js
@@ -1,7 +1,17 @@
+// src/components/Categories.js
 import React from 'react';
+import { useSelector } from 'react-redux';
+import { selectCategoryStatus } from '../redux/categories/categoriesSlice';
 
-const Categories = () => (
-  <h1>Loading................</h1>
-);
+const Categories = () => {
+  const categoryStatus = useSelector(selectCategoryStatus);
+
+  return (
+    <div>
+      <h1>Categories</h1>
+      <p>Status: {categoryStatus}</p>
+    </div>
+  );
+};
 
 export default Categories;
