+++ src/index.js
@@ -1,12 +1,17 @@
+// src/index.js
 import React from 'react';
-import { createRoot } from 'react-dom';
+import ReactDOM from 'react-dom';
 import { Provider } from 'react-redux';
 import reportWebVitals from './reportWebVitals';
+import store from './redux/store/store';
 import App from './App';
-import store from './redux/store';
 import './index.css';
 
-createRoot(document.getElementById('root')).render(
+const rootElement = document.getElementById('root');
+
+const root = ReactDOM.createRoot(rootElement);
+
+root.render(
   <React.StrictMode>
     <Provider store={store}>
       <App />
