--- src/redux/categoriesSlice.js
@@ -1,13 +1,0 @@
-import { createSlice } from '@reduxjs/toolkit';
-
-const initialState = {
-  status: 'Under construction',
-};
-
-const categoriesSlice = createSlice({
-  name: 'categories',
-  initialState,
-  reducers: {},
-});
-
-export default categoriesSlice.reducer;
