+++ src/redux/store/store.js
@@ -1,6 +1,6 @@
 import { configureStore } from '@reduxjs/toolkit';
-import booksReducer from './booksSlice';
-import categoriesReducer from './categoriesSlice';
+import booksReducer from '../book/booksSlice';
+import categoriesReducer from '../categories/categoriesSlice';
 
 const store = configureStore({
   reducer: {
